Productos_Microservicio/app/requirements.txt
requests==2.25.1
fastapi==0.66.0
mysql-connector-python==8.0.26